Making the metricsdb files to keep around configurable.
Currently, the number of most recent metricsdb files to keep around is
determined by a code level constant. In some cases there might be a
requirement to keep more files around (or for the reader to never delete
them). This change introduces a config for that in the
performance-analyzer.properties file named 'max-metrics-db-files'.
Setting this to value -1, makes the files to never be deleted,
otherwise, the acceptable numbers are between 1 and 1000. Anything else
sets the config to default which is 4.

Github issue #47