PA will no longer crash when SecurityManager says no
PA attempts to set the default SSL Socket Factory (which defines rules
for the SSL Sockets it creates) as well as default hostname verification
rules when it is initialized by the Elasticsearch plugin loader.

However, this behavior would result in an AccessControlException when
run alongside the opendistro-security plugin. This commit is a simple
fix which allows these two plugins to work together.