IT improvements (#143)
* Use true/false instead of null/present for integTest props

integTest is a gradle task which runs our integration tests. It uses
system properties like -Dtests.useDockerCluster to decide whether or not
to perform certain actions like spinning up a docker cluster for
testing. The task would previously perform the property's action if the
property was present.

This commit makes the integTest task only execute a system property
action if that property is set to "true"

* Make IT port number configurable

The PerformanceAnalyzerIT class previously assumed that the Performance
Analyzer Webservice would always be listening on port 9600 for any
deployment of PerformanceAnalyzer.

Since this isn't always the case, this commit makes the port number
configurable through a gradle property.

* Allow logging to be enabled for ensurePaAndRcaEnabled